+++ internal/model/vo.go
@@ -1,7 +1,7 @@
 package model
 
 type UserVO struct {
-	ID              uint   `json:"id"`
+	ID              int64  `json:"id"`
 	Name            string `json:"name"`
 	FollowCount     int64  `json:"follow_count"`
 	FollowerCount   int64  `json:"follower_count"`
@@ -26,7 +26,7 @@ func ParseUserVO(user User) UserVO {
 }
 
 type VideoVO struct {
-	ID            uint   `json:"id"`
+	ID            int64  `json:"id"`
 	Author        UserVO `json:"author"`
 	PlayUrl       string `json:"play_url"`
 	CoverUrl      string `json:"cover_url"`
@@ -51,7 +51,7 @@ func ParseVideoVO(video Video) VideoVO {
 }
 
 type MessageVO struct {
-	ID         uint   `json:"id"`
+	ID         int64  `json:"id"`
 	ToUserId   int64  `json:"to_user_id"`
 	FromUserId int64  `json:"from_user_id"`
 	Content    string `json:"content"`
